update login actions
