+++ src/db/seeder.ts
@@ -71,7 +71,7 @@ async function seedDb() {
             dateOfBirth: '1970-01-01',
             roles: ['SYSTEM_ADMINISTRATOR'],
             isLockedOut: false,
-            twoFactorEnabled: false,
+            isTwoFactorEnabled: false,
             twoFactorSecret: null,
             twoFactorTempSecret: null,
             twoFactorRecoveryCodes: null,
